Add function getCloneBaseUrlOrHost in GerritSettings.
When "host" and "close base url" haven't got the same IP, we cannot fetch a commit from gerrit server,
because only the client web url was called.
And this error was notified : "Could not fetch commit because no remote url matches Gerrit host".
Now, if in the settings a "clone base url" is filled, this one will be used for fetching the desired commit.